refactor(schema): move shared serialize and transform helpers to schema root
Sibling subcommands (push, rollback) were cross-importing from each other.
Extract serialize and transform utilities to the shared schema/ parent so
neither subcommand depends on the other.
